Fix node and tank imports
