+++ wsimod/nodes/land.py
@@ -8,8 +8,9 @@ from bisect import bisect_left
 from math import exp, log, log10, sin
 
 from wsimod.core import constants
-from wsimod.nodes.nodes import DecayTank, Node, ResidenceTank
+from wsimod.nodes.nodes import Node
 from wsimod.nodes.nutrient_pool import NutrientPool
+from wsimod.nodes.tanks import DecayTank, ResidenceTank
 
 
 class Land(Node):
