+++ wsimod/nodes/sewer.py
@@ -5,7 +5,8 @@
 Converted to totals on 2022-05-03
 """
 from wsimod.core import constants
-from wsimod.nodes.nodes import Node, QueueTank
+from wsimod.nodes.nodes import Node
+from wsimod.nodes.tanks import QueueTank
 
 
 class Sewer(Node):
