+++ wsimod/nodes/wtw.py
@@ -7,7 +7,8 @@ Converted to totals on 2022-05-03
 from typing import Any, Dict
 
 from wsimod.core import constants
-from wsimod.nodes.nodes import Node, Tank
+from wsimod.nodes.nodes import Node
+from wsimod.nodes.tanks import Tank
 
 
 class WTW(Node):
